pyproject.toml
[project]
name = "ikea-stock-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "duckdb>=1.1.3",
    "firecrawl-py>=1.6.1",
    "orjson>=3.10.12",
    "pandas>=2.2.3",
    "pydantic>=2.10.2",
    "requests>=2.32.3",
    "selenium>=4.27.1",
    "webdriver-manager>=4.0.2",
    "python-dotenv>=1.0.0"
]

[tool.setuptools]
packages = ["ikea_stock_api"]
include-package-data = false
